TLD Fixes & Optimizations
1. TLD now have module structure
2. Made some small code optimizations
3. Fixed Ensemble Classifier according to the original paper - 10
randomized ferns
4. Added comments to most of the functions and methods
